requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
async-timeout==5.0.1
attrs==24.2.0
aws-sam-translator==1.94.0
aws-xray-sdk==2.14.0
bcrypt==4.2.0
bidict==0.23.1
black==24.10.0
blinker==1.8.2
boto3==1.35.84
botocore==1.35.84
certifi==2024.8.30
cffi==1.17.1
cfn-lint==1.20.1
charset-normalizer==3.4.0
click==8.1.7
compose==1.6.2
cryptography==44.0.0
Cython==3.0.11
Deprecated==1.2.15
dnspython==2.7.0
docker==7.1.0
eventlet==0.37.0
exceptiongroup==1.2.2
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
gevent==24.11.1
gevent-websocket==0.10.1
graphql-core==3.2.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joserfc==1.0.0
jsondiff==2.2.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
moto==5.0.21
mpmath==1.3.0
multipart==1.2.0
mypy-extensions==1.0.0
networkx==3.4.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
ordered-set==4.1.0
packaging==24.1
pathable==0.4.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
podman-compose==1.2.0
psycopg2==2.9.10
py-partiql-parser==0.5.6
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
pytest-docker==3.1.1
pytest-docker-tools==3.1.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.25.3
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
setuptools==75.3.0
simple-websocket==1.1.0
six==1.16.0
sympy==1.13.3
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
Werkzeug==3.1.1
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
zope.event==5.0
zope.interface==7.2
